Adds a bunch of temporary assets with some general project house cleaning
Adds various models and sprites to the project for varied uses. Includes
some of the potential UI assets which are temporarily included in the
"movetest_withIndicators" scene.

Attempted to tidy up some of the project files by moving assets and
changing the names of many of the assets to follow a common naming
convention.